No more Airplane mode requirement
Previously it was required for users on some versions of Android to turn off
mobile data in order to set the default route to the Althea device wifi network.

Android will not route traffic over WiFi if that network is not detected to have internet
access, but of course the primary use case for RouterRescue is when the Alhtea device is
offline.

This new patch binds the app to the WiFI interface regardless of the larger system connectivity
state.